pyproject.toml
[project]
name = "022-breakfast-report"
version = "1.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.0",
    "pygelf>=0.4.2",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]

[dependency-groups]
dev = ["pytest>=8.4.1"]
lint = ["ruff>=0.12.0"]

[tool.pytest.ini_options]
pythonpath = "."
